# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A router for generative AI models with OpenAI-compatible API.",
        "title": "Generative API Router",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/aashari/go-generative-api-router",
            "email": "support@yourcompany.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "genapi.example.com",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Returns structured health information including status, services, and version details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check endpoint",
                "responses": {
                    "200": {
                        "description": "Structured health response",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.HealthResponse"
                        }
                    }
                }
            }
        },
        "/v1/chat/completions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Routes chat completion requests to different language model providers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Chat completions API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional vendor to target (e.g., 'openai', 'gemini')",
                        "name": "vendor",
                        "in": "query"
                    },
                    {
                        "description": "Chat completion request in OpenAI-compatible format",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.ChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OpenAI-compatible chat completion response",
                        "schema": {
                            "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.ChatCompletionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "schema": {
                            "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/models": {
            "get": {
                "description": "Returns a list of available language models in OpenAI-compatible format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "List available models",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional vendor to filter models (e.g., 'openai', 'gemini')",
                        "name": "vendor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of available models",
                        "schema": {
                            "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.ModelsResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_aashari_go-generative-api-router_internal_types.ChatCompletionRequest": {
            "type": "object"
        },
        "github_com_aashari_go-generative-api-router_internal_types.ChatCompletionResponse": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.Choice"
                    }
                },
                "created": {
                    "type": "integer",
                    "example": 1677652288
                },
                "id": {
                    "type": "string",
                    "example": "chatcmpl-abc123"
                },
                "model": {
                    "type": "string",
                    "example": "gpt-4o"
                },
                "object": {
                    "type": "string",
                    "example": "chat.completion"
                },
                "service_tier": {
                    "type": "string",
                    "example": "default"
                },
                "system_fingerprint": {
                    "type": "string",
                    "example": "fp_abc123"
                },
                "usage": {
                    "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.Usage"
                }
            }
        },
        "github_com_aashari_go-generative-api-router_internal_types.Choice": {
            "type": "object",
            "properties": {
                "finish_reason": {
                    "type": "string",
                    "example": "stop"
                },
                "index": {
                    "type": "integer",
                    "example": 0
                },
                "logprobs": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.Message"
                }
            }
        },
        "github_com_aashari_go-generative-api-router_internal_types.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "invalid_model"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid model specified"
                },
                "param": {
                    "type": "string",
                    "example": "model"
                },
                "type": {
                    "type": "string",
                    "example": "invalid_request_error"
                }
            }
        },
        "github_com_aashari_go-generative-api-router_internal_types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.ErrorInfo"
                }
            }
        },
        "github_com_aashari_go-generative-api-router_internal_types.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Hello, how are you?"
                },
                "name": {
                    "type": "string",
                    "example": "John"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                },
                "tool_call_id": {
                    "type": "string"
                },
                "tool_calls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.ToolCall"
                    }
                }
            }
        },
        "github_com_aashari_go-generative-api-router_internal_types.Model": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer",
                    "example": 1677610602
                },
                "id": {
                    "type": "string",
                    "example": "gpt-4o"
                },
                "object": {
                    "type": "string",
                    "example": "model"
                },
                "owned_by": {
                    "type": "string",
                    "example": "openai"
                }
            }
        },
        "github_com_aashari_go-generative-api-router_internal_types.ModelsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.Model"
                    }
                },
                "object": {
                    "type": "string",
                    "example": "list"
                }
            }
        },
        "github_com_aashari_go-generative-api-router_internal_types.TokenDetails": {
            "type": "object",
            "properties": {
                "accepted_prediction_tokens": {
                    "type": "integer",
                    "example": 0
                },
                "audio_tokens": {
                    "type": "integer",
                    "example": 0
                },
                "cached_tokens": {
                    "type": "integer",
                    "example": 0
                },
                "reasoning_tokens": {
                    "type": "integer",
                    "example": 0
                },
                "rejected_prediction_tokens": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "github_com_aashari_go-generative-api-router_internal_types.ToolCall": {
            "type": "object",
            "properties": {
                "function": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string",
                    "example": "call_8qty38"
                },
                "type": {
                    "type": "string",
                    "example": "function"
                }
            }
        },
        "github_com_aashari_go-generative-api-router_internal_types.Usage": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer",
                    "example": 20
                },
                "completion_tokens_details": {
                    "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.TokenDetails"
                },
                "prompt_tokens": {
                    "type": "integer",
                    "example": 10
                },
                "prompt_tokens_details": {
                    "description": "Added additional usage details for OpenAI compatibility",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_aashari_go-generative-api-router_internal_types.TokenDetails"
                        }
                    ]
                },
                "total_tokens": {
                    "type": "integer",
                    "example": 30
                }
            }
        },
        "internal_handlers.HealthResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "services": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the API key value.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}